+++ src/Contracts/SearchDialect.php
@@ -9,6 +9,8 @@ use Illuminate\Support\Collection;
 
 interface SearchDialect
 {
+    public function getName(): string;
+
     public function buildQueries(Collection $modelsToSearchThrough): Collection;
 
     public function addWhereTermsToQuery(Builder $query, array|string $column): void;
@@ -16,4 +18,12 @@ interface SearchDialect
     public function useSoundex(): void;
 
     public function avoidSoundex(): void;
+
+    public function makeCoalesce(Collection $keys): string;
+
+    public function charLengthFunction(): string;
+
+    public function supportsFullTextSearch(): bool;
+
+    public function supportsOrderByModel(): bool;
 }